requirements.txt
progress==1.6
discord==2.1.0; python_version >= "3.6"
PlexAPI==4.*
tautulli==3.1.3.2103
SQLAlchemy-Utils==0.38.3
SQLAlchemy==1.4.40
trakt==3.4.0
IMDbPY==2022.7.9
asyncio==3.4.3
confuse~=1.7
